+++ src/main/java/kitchenpos/menu/application/MenuResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.application.dto;
+package kitchenpos.menu.application;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuResponse {
