+++ src/main/java/kitchenpos/menu/domain/MenuDao.java
@@ -1,8 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.Menu;
 
 public interface MenuDao {
     Menu save(Menu entity);
