+++ src/main/java/kitchenpos/menu/domain/MenuProductDao.java
@@ -1,8 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.MenuProduct;
 
 public interface MenuProductDao {
     MenuProduct save(MenuProduct entity);
