+++ src/main/java/kitchenpos/menu/domain/infra/JdbcTemplateMenuDao.java
@@ -1,11 +1,11 @@
-package kitchenpos.dao.infrastructure;
+package kitchenpos.menu.domain.infra;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
