+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -9,14 +9,14 @@ import static kitchenpos.Fixture.테이블;
 import static kitchenpos.Fixture.테이블집합;
 
 import java.math.BigDecimal;
-import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.TableGroupDao;
-import kitchenpos.domain.Menu;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.domain.MenuGroupDao;
 import kitchenpos.product.domain.Product;
